Always store implicit output and add Value.LongValue

+++ MiniScript-cs/MiniscriptIntrinsics.cs
@@ -312,7 +312,7 @@ namespace Miniscript {
 			if (targetType == typeof(double)) return value == null ? 0.0 : value.DoubleValue();
 			if (targetType == typeof(float)) return value == null ? 0f : value.FloatValue();
 			if (targetType == typeof(int)) return value == null ? 0 : value.IntValue();
-			if (targetType == typeof(long)) return value == null ? 0L : (long)value.DoubleValue();
+			if (targetType == typeof(long)) return value == null ? 0L : value.LongValue();
 			if (targetType == typeof(uint)) return value == null ? 0U : value.UIntValue();
 			if (targetType == typeof(ulong)) return value == null ? 0UL : (ulong)value.DoubleValue();
 			if (targetType == typeof(bool)) return value != null && value.BoolValue();
